# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "license": {
      "name": "Proprietary"
    },
    "title": "School Guardian — Support Platform API",
    "version": "0.1.0",
    "description": "API de gerenciamento de tickets da School Guardian.\n\n## Autenticação\n1. `POST /auth/login` com e-mail e senha retorna um Bearer token válido por **10 minutos**.\n2. Envie `Authorization: Bearer <token>` em todas as rotas protegidas.\n3. **Toda resposta autenticada retorna um token renovado no header `X-Access-Token`** (e `X-Access-Token-Expires-In`). Substitua o token armazenado por esse valor a cada resposta.\n4. A sessão expira após 10 min sem uso ou ao atingir o tempo máximo (padrão 8h desde o login).\n5. `POST /auth/logout`, troca de senha e desativação revogam todos os tokens do usuário.\n\n## Erros\nTodas as respostas de erro seguem o formato `{ statusCode, code, message, details? }`."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TicketOptions": {
        "type": "object",
        "properties": {
          "status": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "value",
                "label"
              ],
              "additionalProperties": false
            }
          },
          "devStatus": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "value",
                "label"
              ],
              "additionalProperties": false
            }
          },
          "entryType": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "value",
                "label"
              ],
              "additionalProperties": false
            }
          },
          "slaType": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "value": {
                  "type": "string"
                },
                "label": {
                  "type": "string"
                }
              },
              "required": [
                "value",
                "label"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "status",
          "devStatus",
          "entryType",
          "slaType"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/tickets/options": {
      "get": {
        "operationId": "getTicketOptions",
        "summary": "Valores e rótulos dos campos de domínio do ticket",
        "tags": [
          "tickets"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "headers": {
              "x-access-token": {
                "schema": {
                  "type": "string"
                },
                "description": "Novo Bearer token (validade renovada). O cliente deve substituir o token armazenado por este valor."
              },
              "x-access-token-expires-in": {
                "schema": {
                  "type": "string",
                  "example": "600"
                },
                "description": "Validade do novo token, em segundos."
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketOptions"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido, expirado ou revogado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Usuário sem permissão para este recurso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "code",
                    "message"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "field",
                          "message"
                        ],
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local (npm run dev)"
    },
    {
      "url": "https://{apiId}.execute-api.{region}.amazonaws.com/{stage}",
      "description": "AWS API Gateway",
      "variables": {
        "apiId": {
          "default": "api-id"
        },
        "region": {
          "default": "us-east-1"
        },
        "stage": {
          "default": "dev",
          "enum": [
            "dev",
            "staging",
            "prod"
          ]
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}